backend/requirements.txt
# Internal trace:
# - Wrong before: the backend directory lacked its own requirements file even though it is started independently.
# - Fixed now: backend installs can use this colocated manifest directly.

fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic-settings>=2.4.0
transformers>=4.45.0
torch>=2.4.0
torchvision>=0.19.0
timm>=1.0.9
opencv-python-headless>=4.10.0.84
moviepy>=1.0.3
ffmpeg-python>=0.2.0
pillow>=10.4.0
numpy>=1.26.4
scipy>=1.14.1
soundfile>=0.12.1
librosa>=0.10.2.post1
aiofiles>=24.1.0
httpx>=0.27.0
pytest>=8.3.0
